move the last flagset tests out of main
